backend/requirements.txt
asgiref==3.8.1
Django==5.2.2
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.10.1
python-dateutil==2.9.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
gunicorn==23.0.0
